Added logs and moved files to correct folders
