+++ backend/users/src/app/resolvers/applicationSearch.resolver.ts
@@ -1,7 +1,7 @@
 import { Resolver, Query, Args, Int } from '@nestjs/graphql';
 import { ApplicationSearchService } from '../services/applicationSearch.service';
 import { ApplicationSearchResponse } from '../dto/reponse/applicationSearchResponse';
-import { ApplicationFilter } from '../dto/applicationFilter.enum';
+import { ApplicationFilter } from '../utilities/enums/applicationFilter.enum';
 
 @Resolver()
 export class ApplicationSearchResolver {
