+++ backend/users/src/app/services/application/appParticipant.service.spec.ts
@@ -8,7 +8,8 @@ import { getRepositoryToken } from '@nestjs/typeorm';
 
 import { Repository } from 'typeorm';
 import { AppParticipant } from '../../entities/appParticipant.entity';
-import { ViewAppParticipantsDto } from '../../dto/appParticipants/viewAppParticipantsDto';
+import { ViewAppParticipantsDto } from '../../dto/appParticipants/viewAppParticipants.dto';
+import { AppParticipantFilter } from 'src/app/dto/appParticipants/appParticipantFilter.enum';
 
 describe('AppParticipantsService', () => {
   let service: AppParticipantService;
@@ -62,7 +63,7 @@ describe('AppParticipantsService', () => {
         ),
       );
 
-      const result = await service.getAppParticipantsByAppId(1, 'user');
+      const result = await service.getAppParticipantsByAppId(1, 'user', AppParticipantFilter.MAIN);
       expect(result).toEqual(
         plainToInstance(ViewAppParticipantsDto, [
           {
@@ -85,7 +86,7 @@ describe('AppParticipantsService', () => {
         });
 
       await expect(
-        service.getAppParticipantsByAppId(1, 'user'),
+        service.getAppParticipantsByAppId(1, 'user', AppParticipantFilter.MAIN),
       ).rejects.toThrow(
         new HttpException(
           'Failed to retrieve app participants by appId: 1',
