+++ backend/users/src/app/utilities/enums/applicationFilter.enum.ts
