+++ cats-frontend/src/app/features/applications/application/applicationTabs/appParticipants/graphql/Participants.graphql
